requirements.txt
django>=4.2,<5.0
django-ipware
webauthn
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.11
dj-database-url==2.1.0
requests
gradio_client>=1.3.0
huggingface_hub
python-dotenv
google-genai
reportlab